Address warnings
With the previous update, additional warnings are generated by default.
This patch addresses these:

 * -Wunused-function: use `inline` instead of `static` in header.
 * -Wunused-but-set-variable: code is protected in #ifndef NDEBUG when
   it is only used for the purpose of assertions.
 * -Wunused-function and -Wunused-but-set-variable: dead code is
   removed.
 * -Wsign-compare: uses of `auto` are replaced with `unsigned` or
   `size_t`, or use `static_cast`.
 * -Wimplicit-fallthrough: this warning is silenced using special
   comments.
 * -Wpedantic: remove extra `;` on closing brace for namespace.
 * -Woverflow: use types explicitly sized (e.g. std::uint64_t) and
   appropriate macros for minimum-width integer constants instead of
   size_t.

+++ lib/Builtins.cpp
@@ -57,7 +57,7 @@ std::string GetUnmangledName(const std::string &str, size_t *pos) {
     return "";
   }
   ptrdiff_t name_pos = end - str.data();
-  if (name_pos + name_len > str.size()) {
+  if (static_cast<std::size_t>(name_pos + name_len) > str.size()) {
     // Protect against maliciously large number.
     return "";
   }
